+++ lib/hl7/message.rb
