+++ lib/hl7/message_parser.rb
