+++ lib/hl7/segment_default.rb
