+++ lib/hl7/segment_fields.rb
