+++ lib/hl7/segment_generator.rb
