+++ lib/hl7/segment_list_storage.rb
