+++ lib/hl7/segments/err.rb
