+++ lib/hl7/segments/ft1.rb
