+++ lib/hl7/segments/nte.rb
