+++ lib/hl7/segments/rf1.rb
